🐛 Count is an integer, what else can it be?

+++ kmserver-core/src/dao/sqls/kara.ts
@@ -29,7 +29,7 @@ export const getAllKaras = (
 WITH ${withCTE.join(', \n')}
 SELECT
 ${onlyCount ? `
-  count(ak.pk_kid) AS count`
+  count(ak.pk_kid)::integer AS count`
 : `
   ak.tags AS tags,
   ak.pk_kid AS kid,
